Fix tests, modules, CLAUDE, markdown file width
