--- crates/selfie/src/commands/tests.rs
@@ -1,210 +1,0 @@
-//! Test modules for command functionality
-
-// Error context tests temporarily disabled due to compilation issues
-// mod error_context_tests;
-
-// Simple tests demonstrating enhanced CommandError context functionality
-
-use std::path::PathBuf;
-use std::sync::Arc;
-use std::time::Duration;
-
-use crate::commands::runner::CommandError;
-
-#[test]
-fn test_timeout_error_contains_rich_context() {
-    let error = CommandError::Timeout {
-        command: "sleep 10".to_string(),
-        timeout: Duration::from_secs(5),
-        working_directory: PathBuf::from("/home/user/project"),
-    };
-
-    // Test error message content
-    let error_message = error.to_string();
-    assert!(error_message.contains("sleep 10"));
-    assert!(error_message.contains("5s"));
-
-    // Test that context fields are accessible
-    match error {
-        CommandError::Timeout {
-            command,
-            timeout,
-            working_directory,
-        } => {
-            assert_eq!(command, "sleep 10");
-            assert_eq!(timeout, Duration::from_secs(5));
-            assert_eq!(working_directory, PathBuf::from("/home/user/project"));
-        }
-        _ => panic!("Expected Timeout error"),
-    }
-}
-
-#[test]
-fn test_io_error_contains_command_context() {
-    let io_error = std::io::Error::new(std::io::ErrorKind::NotFound, "command not found");
-    let error = CommandError::IoError {
-        command: "nonexistent-command".to_string(),
-        working_directory: PathBuf::from("/tmp"),
-        source: Arc::new(io_error),
-    };
-
-    // Test error message content
-    let error_message = error.to_string();
-    assert!(error_message.contains("nonexistent-command"));
-
-    // Test that context fields are accessible
-    match error {
-        CommandError::IoError {
-            command,
-            working_directory,
-            source,
-        } => {
-            assert_eq!(command, "nonexistent-command");
-            assert_eq!(working_directory, PathBuf::from("/tmp"));
-            assert_eq!(source.kind(), std::io::ErrorKind::NotFound);
-        }
-        _ => panic!("Expected IoError"),
-    }
-}
-
-#[test]
-fn test_non_zero_exit_error_contains_full_context() {
-    let error = CommandError::NonZeroExit {
-        command: "false".to_string(),
-        exit_code: 1,
-        stdout: "".to_string(),
-        stderr: "Command failed".to_string(),
-        working_directory: PathBuf::from("/home/user"),
-        execution_duration: Duration::from_millis(500),
-    };
-
-    // Test error message content
-    let error_message = error.to_string();
-    assert!(error_message.contains("false"));
-    assert!(error_message.contains("exit code 1"));
-
-    // Test that all context is accessible
-    match error {
-        CommandError::NonZeroExit {
-            command,
-            exit_code,
-            stdout,
-            stderr,
-            working_directory,
-            execution_duration,
-        } => {
-            assert_eq!(command, "false");
-            assert_eq!(exit_code, 1);
-            assert_eq!(stdout, "");
-            assert_eq!(stderr, "Command failed");
-            assert_eq!(working_directory, PathBuf::from("/home/user"));
-            assert_eq!(execution_duration, Duration::from_millis(500));
-        }
-        _ => panic!("Expected NonZeroExit error"),
-    }
-}
-
-#[test]
-fn test_command_error_debug_output_includes_context() {
-    let error = CommandError::Timeout {
-        command: "debug-test-command".to_string(),
-        timeout: Duration::from_secs(30),
-        working_directory: PathBuf::from("/debug/test"),
-    };
-
-    let debug_output = format!("{error:?}");
-    assert!(debug_output.contains("debug-test-command"));
-    assert!(debug_output.contains("30s"));
-    assert!(debug_output.contains("/debug/test"));
-}
-
-#[test]
-fn test_command_error_source_chain() {
-    let original_error =
-        std::io::Error::new(std::io::ErrorKind::PermissionDenied, "Permission denied");
-    let error = CommandError::IoError {
-        command: "restricted-command".to_string(),
-        working_directory: PathBuf::from("/restricted"),
-        source: Arc::new(original_error),
-    };
-
-    // Test that we can access the source error through the Arc
-    let source = std::error::Error::source(&error);
-    assert!(source.is_some());
-
-    // The source is the Arc<std::io::Error>, so we need to dereference it
-    let arc_error = source
-        .unwrap()
-        .downcast_ref::<Arc<std::io::Error>>()
-        .unwrap();
-    assert_eq!(arc_error.kind(), std::io::ErrorKind::PermissionDenied);
-}
-
-#[test]
-fn test_command_errors_are_cloneable() {
-    let timeout_error = CommandError::Timeout {
-        command: "test".to_string(),
-        timeout: Duration::from_secs(1),
-        working_directory: PathBuf::from("/test"),
-    };
-
-    let cloned_error = timeout_error.clone();
-    assert_eq!(format!("{timeout_error}"), format!("{cloned_error}"));
-}
-
-#[test]
-fn test_command_error_context_extraction_for_debugging() {
-    let working_dir = PathBuf::from("/my/working/directory");
-    let command = "my-test-command --verbose";
-    let timeout = Duration::from_secs(30);
-
-    let error = CommandError::Timeout {
-        command: command.to_string(),
-        timeout,
-        working_directory: working_dir.clone(),
-    };
-
-    // Demonstrate extracting context for logging/debugging
-    let debug_info = match &error {
-        CommandError::Timeout {
-            command: cmd,
-            timeout: t,
-            working_directory: wd,
-        } => {
-            format!(
-                "Command '{}' timed out after {:?} in directory '{}'",
-                cmd,
-                t,
-                wd.display()
-            )
-        }
-        _ => panic!("Expected Timeout error"),
-    };
-
-    assert!(debug_info.contains(command));
-    assert!(debug_info.contains("30s"));
-    assert!(debug_info.contains("/my/working/directory"));
-}
-
-#[test]
-fn test_io_error_preserves_original_error_information() {
-    let original_error =
-        std::io::Error::new(std::io::ErrorKind::BrokenPipe, "The pipe has been broken");
-    let original_kind = original_error.kind();
-    let original_message = original_error.to_string();
-
-    let command_error = CommandError::IoError {
-        command: "pipe-command".to_string(),
-        working_directory: PathBuf::from("/pipes"),
-        source: Arc::new(original_error),
-    };
-
-    // Verify the original error information is preserved
-    match command_error {
-        CommandError::IoError { source, .. } => {
-            assert_eq!(source.kind(), original_kind);
-            assert_eq!(source.to_string(), original_message);
-        }
-        _ => panic!("Expected IoError"),
-    }
-}
